pyproject.toml
[tool.poetry]
name = "ocr-analysis"
version = "0.1.0"
description = "optical character recognition analysis"
authors = ["Buğra Çayır"]
readme = "README.md"
packages = [{include = "ocr_analysis"}]

[tool.poetry.dependencies]
python = ">=3.8, <3.11.9"
redis = "^5.0.0"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
dynaconf = "^3.2.1"
jinja2 = "^3.1.2"
black = "^23.7.0"
bandit = "^1.7.5"
isort = "^5.12.0"
python-multipart = "^0.0.6"
pytesseract = "^0.3.10"
pillow = "^10.0.0"
validators = "^0.21.2"
python-whois = "^0.8.0"
requests = "^2.31.0"
dateparser = "^1.1.8"
textract = "^1.6.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
